remove aliasing and align BEE behaviour with targets
